fix: Change dashboard avatar to use user's real avatar
